Introduce `empty_as_none` and `empty_as_none_vec` deserializers
This commit changes the way `foo=` is deserialized into `Option<_>`. Previously it was deserialized as `None`, now it will be deserialized as `Some(_)`, unless the new `empty_as_none` deserializer is used via `#[serde(deserialize_with = "crate::de::empty_as_none")]`. This allows users to decide on the deserialization strategy on a field-by-field basis.