Implement optional "empty is `Some`" behavior
This commit introduces an internal const generic `EMPTY_IS_SOME` parameter, that determines if empty strings should be treated as `None` or `Some("")`. The behavior of `from_bytes()` and `from_str()` stays the same, but an additional `empty_is_some` module with the same fns is introduced, that allows users to choose the alternative behavior.